fix(storage): align provider rename boundaries
