+++ packages/platform/service-platform/src/file-storage-nativefs.ts
@@ -20,6 +20,7 @@ import type {
   FileStorageChangeEvent,
 } from '@bangle.io/types';
 import { isVisibleWorkspaceDirectoryName, WsPath } from '@bangle.io/ws-path';
+import { assertSameWorkspaceRename } from './file-storage-utils';
 
 type Config = {
   getRootDirHandle: (
@@ -247,19 +248,7 @@ export class FileStorageNativeFs
     },
   ): Promise<void> {
     await this.mountPromise;
-    const oldPath = WsPath.assertFile(wsPath);
-    const newPath = WsPath.assertFile(newWsPath);
-    if (oldPath.wsName !== newPath.wsName) {
-      throwAppError(
-        'error::file:invalid-operation',
-        'Cannot rename file across different workspaces',
-        {
-          operation: 'rename',
-          oldWsPath: wsPath,
-          newWsPath,
-        },
-      );
-    }
+    assertSameWorkspaceRename(wsPath, newWsPath);
     const fs = await this.getFs({ wsPath });
     try {
       await fs.moveFile(
